requirements.txt
# Dépendances pour le script d'automatisation quotidienne
notion-client>=2.7.0
python-dotenv>=1.2.1
python-dateutil>=2.9.0

# Dépendances pour le MCP Veille Thèses
mcp>=1.0.0
httpx>=0.25.0
beautifulsoup4>=4.12.0
pyyaml>=6.0
lxml>=4.9.0
aiohttp>=3.9.0
